+++ cmd/hidden_lake/applications/messenger/internal/database/database_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/number571/go-peer/pkg/crypto/asymmetric"
-	"github.com/number571/go-peer/pkg/storage/database"
 	testutils "github.com/number571/go-peer/test/utils"
 )
 
@@ -31,9 +30,7 @@ func TestDatabase(t *testing.T) {
 	os.RemoveAll(tcPath)
 	defer os.RemoveAll(tcPath)
 
-	db, err := NewKeyValueDB(database.NewSettings(&database.SSettings{
-		FPath: tcPath,
-	}))
+	db, err := NewKeyValueDB(tcPath)
 	if err != nil {
 		t.Error(err)
 		return