+++ cmd/hidden_lake/applications/messenger/pkg/app/init_database.go
@@ -6,15 +6,10 @@ import (
 
 	hlm_database "github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/internal/database"
 	hlm_settings "github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/pkg/settings"
-	"github.com/number571/go-peer/pkg/storage/database"
 )
 
 func (p *sApp) initDatabase() error {
-	db, err := hlm_database.NewKeyValueDB(
-		database.NewSettings(&database.SSettings{
-			FPath: filepath.Join(p.fPathTo, hlm_settings.CPathDB),
-		}),
-	)
+	db, err := hlm_database.NewKeyValueDB(filepath.Join(p.fPathTo, hlm_settings.CPathDB))
 	if err != nil {
 		return fmt.Errorf("open KV database: %w", err)
 	}