+++ cmd/hidden_lake/service/pkg/app/init_anon_node.go
@@ -41,11 +41,7 @@ func (p *sApp) initAnonNode() error {
 		cfgSettings = cfg.GetSettings()
 	)
 
-	kvDatabase, err := database.NewKVDatabase(
-		database.NewSettings(&database.SSettings{
-			FPath: filepath.Join(p.fPathTo, hls_settings.CPathDB),
-		}),
-	)
+	kvDatabase, err := database.NewKVDatabase(filepath.Join(p.fPathTo, hls_settings.CPathDB))
 	if err != nil {
 		return utils.MergeErrors(ErrOpenKVDatabase, err)
 	}