+++ pkg/network/anonymity/anonymity_test.go
@@ -665,13 +665,7 @@ func (p *stLogging) HasErro() bool {
 */
 
 func testNewNode(timeWait time.Duration, addr string, typeDB, numDB int) (INode, context.CancelFunc) {
-	db, err := database.NewKVDatabase(
-		database.NewSettings(&database.SSettings{
-			FPath:     fmt.Sprintf(tcPathDBTemplate, typeDB, numDB),
-			FWorkSize: testutils.TCWorkSize,
-			FPassword: "CIPHER",
-		}),
-	)
+	db, err := database.NewKVDatabase(fmt.Sprintf(tcPathDBTemplate, typeDB, numDB))
 	if err != nil {
 		panic(err)
 	}