+++ pkg/network/anonymity/examples/echo/construct.go
@@ -63,11 +63,7 @@ func newNode(serviceName, address string) anonymity.INode {
 			},
 		),
 		func() database.IKVDatabase {
-			db, err := database.NewKVDatabase(
-				database.NewSettings(&database.SSettings{
-					FPath: "./database_" + serviceName + ".db",
-				}),
-			)
+			db, err := database.NewKVDatabase("./database_" + serviceName + ".db")
 			if err != nil {
 				panic(err)
 			}