+++ pkg/storage/database/types.go
@@ -6,15 +6,8 @@ import (
 
 type IKVDatabase interface {
 	types.ICloser
-	GetSettings() ISettings
 
 	Set([]byte, []byte) error
 	Get([]byte) ([]byte, error)
 	Del([]byte) error
-}
-
-type ISettings interface {
-	GetPath() string
-	GetWorkSize() uint64
-	GetPassword() string
 }